moving jira issues
